requirements.txt
matplotlib
numpy
pandas
tweepy
Pillow
wordcloud
nltk
textblob
pycountry
scikit-learn
langdetect
